requirements.txt
biopython
lxml
BeautifulSoup4
html5lib
luigi
pandas
networkx
deepmol[all]
retry
IPython
pytorch-lightning
transformers